Everything is working and formatting EXCEPT adding new course.
The courses table has more columns than the employees table.  Given the extreme similarity in the code for employees versus courses, I am confident that this mismatch between course VALUES( - ) and employee VALUES(-) is the root of the problem.    Secondly, I have labeled three columns in the courses table as BOOLEAN values but I don't know how to test whether or not I have done this correctly.  I went back to UPDATE the table to make them TRUE or FALSE instead of 'TRUE' or 'FALSE' to ensure that they are treated as BOOLEAN operators and not Strings.